+ HttpException

+++ src/product/product.controller.ts
@@ -1,6 +1,7 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Req, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpException, HttpStatus, Param, Post, Put, Req, UseFilters, UseGuards } from '@nestjs/common';
 import { ApiCookieAuth, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { TypeormExceptionFilter } from 'src/typeormException.filter';
 import { CreateProductDto } from './dto/createProduct.dto';
 import { UpdateProductDto } from './dto/updateProduct.dto';
 import { Product } from './entities/product.entity';
@@ -9,19 +10,20 @@ import { ProductService } from './product.service';
 
 @ApiTags('product')
 @Controller('product')
+//@UseFilters(TypeormExceptionFilter)
 export class ProductController {
   constructor(private readonly productService: ProductService) {}
 
   @ApiOperation({summary: 'Получить все товары'})
-  @ApiOkResponse({description: 'getAll', type: Product})
+  @ApiOkResponse({description: 'getAll', type: [Product]})
   @Get()
   getAll() {
     return this.productService.findAll();
   }
 
   @ApiOperation({summary: 'Получить товар по id'})
-  @ApiOkResponse({description: 'getById', type: [Product]})
-  @Get('id')
+  @ApiOkResponse({description: 'getById', type: Product})
+  @Get(':id')
   getById(@Param('id') id: number) {
     return this.productService.FindById(id);
   }
@@ -39,10 +41,17 @@ export class ProductController {
   @ApiOkResponse({description: 'delete'})
   @ApiCookieAuth()
   @UseGuards(JwtAuthGuard)
-  @UseGuards(ProductGuard)
   @Delete(':id')
-  delete(@Param('id') id: number) {
-    this.productService.deleteProduct(id);
+  async delete(@Param('id') id: number, @Req() req) {
+    await this.productService.checkOwner(id, req.user.userId)
+    .then(()=>{
+      this.productService.deleteProduct(id);
+    })
+    .catch((err)=> {
+      throw new HttpException({
+        message: 'not owner.'
+      }, HttpStatus.BAD_REQUEST);
+    });
   }
 
   @ApiOperation({summary: 'Обновить товар'})
@@ -51,8 +60,16 @@ export class ProductController {
   @UseGuards(JwtAuthGuard)
   @UseGuards(ProductGuard)
   @Put(':id') 
-  update(@Param('id') id: number, @Body() shop: UpdateProductDto) {
-    this.productService.update(id, shop);
+  async update(@Param('id') id: number, @Body() shop: UpdateProductDto, @Req() req) {
+    await this.productService.checkOwner(id, req.user.userId)
+    .then(()=>{
+      this.productService.update(id, shop);
+    })
+    .catch((err)=> {
+      throw new HttpException({
+        message: 'not owner.'
+      }, HttpStatus.BAD_REQUEST);
+    });
   }
 
   @ApiOperation({summary: 'Купить товар'})
